adding the sum logic using recursion with java

+++ src/Recursion/FactorialCalculation.java
@@ -2,7 +2,7 @@ package Recursion;
 
 public class FactorialCalculation {
 
-    public long calculateFactorial(int n){
+    public long calculateFactorial(long n){
         if(n==1){
             return 1;
         }else{
